feat: 스토리북 모달 레이아웃 추가
